Major changes in DSL. Closes #177
This commit is a big one, it was quite impossible to split it easily.

What is brings is a new syntax in the DSL that allows for better separation
between gatling-core and gatling-http, which, in the future, will be good for
the other modules as well.

some methods' names were simplified, but the biggest change is the exec method, ie:
 scenario('name')
  .exec( http('request name').get('url')..[conf of request]..capture(regexp(...) in '', xpath(...) in '').check(xpathEquals(...)) )

Using this syntax it is now clear that we can execute an action, whether it's an HTTP request
or a LDAP one.

Little simplifications were made (classes and objects not nested anymore, build doesn't need a scenarioId, etc.)