Allows nested loops. Closes #184
Loops can now be nested.
If the user wants to have access to the counter of a loop, he must declare
his intention thanks to the DSL, eg:
  loop(chain).counterName(myCounter).times(20)
then the counter is available inside the loop, eg:
  exec( (c: Context) => println( c.getCounterValue(myCounter) )
The value is accessible anywhere from inside the loop, if the user tries to access it
from outside the corresponding loop, an Exception will be thrown.