Gatling scenario recorder, the entry point is GatlingHttpProxyUI #13
The ConfigurationFrame is used to specify different kind of settings (port, filters, etc...) for the recorder. Closes #201
RunningFrame represents the running state of the recorder. Closes #202
Forwarded requests are shown in the RunningFrame. Closes #203
The URL filter for Java regular expression is configurable in the ConfigurationFrame. Closes #204
Same as Java regular expression, but for Ant pattern. Closes #205
Preferences can be saved if selected in the ConfigurationFrame. They are saved in the user home directory. Closes #206
Requests with body are automaticaly dumped in a specific directory. Closes #207