Allow `runDescription` to be specified via CLI or conf file
Prior to this commit, if you ran Gatling with the CLI or config
file option for the simulation class defined, it would automatically
set the run description to the same value as the run ID.

This commit changes the behavior to look and see if the user
specified a value for `runDescription` via the CLI or in
the config file, and uses the value if they did.  It falls
back to the old behavior if the user did not specify a value.