Make sure Session.markAsFailed propagates status to groups, close #3197
Motivation:

Currently, only failing checks update group status.
As a consequence, failing to build a request doesn't fail the group, while it's clearly an error that should be reported.

Modifications:

Update all the groups in the block stack with the new status up to the first tryMax block.

Result:

Session status update are properly reported in groups