Use dispatcher instead of scheduler to protect against spins, close #3223
Motivation:

Loop has a protection against spins that would lock the thread.

When a possible spin is detected, we currently schedule the next action
in 1 millisecond.

Unfortunately, LARS is not that precise, and we might end up losing 30
milliseconds instead.

Modification:

Use the ActorSystem’s dispatcher instead.

Result:

Way better performance when hitting a possible spin.