Don't fallback to interactive mode when user provided simulation is invalid, close #3284
Motivation:

When user provides an explicit simulation class name but it's invalid,
we shouldn't fallback to interactive mode. There's a very good chance
the caller is a program and not a human.

Modifications:

Immediately and always throw an exception when user provided class is
invalid (not found, cannot be loaded, not a Simulation).

Result:

Better user experience + don’t stall when launcher is not a human.