Use insecure TrustManager by default, close #3325
Motivation:

Gatling is a load test tool, not a security audit tool. Moreover,
people tend to use a broken TLS configuration for their testing
environment.

The default should be to trust all server certificates, and we should
provide the ability to change the behavior.

Modifications:

* If no TrustManager is specified, use by default an insecure one
* Introduce a config prop to disable insecure TrustManager.

Result:

It’s no longer required to configure a TrustManager when configuring a
custom KeyManager