Introduce a feature for reading feeder records, close #3384
Motivation:

It seems many people have been using RecordSeqFeederBuilder#records as
a hack for reusing our parsers for custom needs.
The records fields was public as RecordSeqFeederBuilder was a case
class, but it wasn't meant to be a public API that people could use. As
a result, it went away when we introduced batched feeders. Obviously,
the whole data is not loaded in memory in the object returned by csv().

Modifications:

Introduce a readRecords method that returns a Seq[Map[String, Any]].

Note that the behavior is different in the sense that the underlying
source will be parsed on each call.

Result:

Convenient API for grabbing feeder records.