Don't pile up updates when redirecting and request is not a resource, close #3386
Motivation:

When performing a redirect, we always pile up updates. This is wrong
when the request is not a resource and will cause updates to be applied
twice, for example account twice the response time in the group’s
cumulatedResponseTime.

Modification:

On redirect, only pile up updates when in a resource.

Result:

Correct cumulatedResponseTime