Automatically parse String values stored in Session into expected type, close #3395
Motivation:

Currently, if an attribute value is a String representation of a
number, trying to extract it from the Session with a number type will
fail.

This is perfectly fine from a type perspective, and is intended to
avoid users parsing Strings over and over again, which would hurt
performance.

Then, this is not intuitive for users who are not used to strongly
typed languages.
Let's relax this behavior and automatically parse Strings if possible.

Modifications:

Implement TypeCasters for well know types so they parse Strings

Result:

More intuitive behavior for users who don’t come from strongly typed
languages.