Stop using unsafeChars in RichCharSequence#indexOf
Motivation:

unsafeChars don’t work with Java 9.
Using it in RichCharSequence#indexOf is completely overkill anyway as
it’s only used in ElCompiler.

Modification:

Take a char array instead and have a constant for those chars in the
one single caller of this method.

Result:

Less dependency on code that no longer works with Java 9