Warmup request should set Connection header to close #3437
Motivation:

Warmup request currently sets Connection header to keep-alive, which is
both useless as it's the default value for HTTP/1.1, and non sense as
we won't be reusing the connection ever.

This connection currently stays open in the pool until it gets closed
for being idle.

Modification:

Set Connection header to close

Result:

No more idle connection.