Release async resources semaphore on redirect to different domain, close #3449
Motivation:

We use a kind of semaphore strategy for capping connections per domain
when fetching resources.
When fetching a resource involves a redirect, we fail to release the
semaphore because we don't keep track of the original domain.
This causes run to hang forever.

Modification:

Keep track of original domain so we properly release semaphore.

Result:

No more test hanging