Optimize default HTTP check
Motivation:

Some parts of RequestBuilder#DefaultHttpCheck are still dynamic and
need to be recomputed over and over again. Also, it performs an
`in(Seq[Int])` which could be optimized.

Modification:

Use a custom validator whose name is static and that uses a BitSet
instead of search in a Seq[Int].

Result:

Faster default HTTP check