Gatling "wait" method clashes with Object#wait
Motivation:

We want to be able to pass a simple int to wait and get it implicitly lifted into FiniteDuration.
Sadly, this doesn't work because of Java's Object#wait(long).

Modification:

Rename "wait" into "await" in WebSocket and SSE support.

Result:

It's now possible to pass an Int to await without duration unit.