Compute most of request headers in RequestBuilder
Motivation:

Headers are patched in WritableRequestBuilder, which is very late.

Modification:

Patch most of the headers upstream in RequestBuilder.
The only header we can't patch there is Content-Length as we only the compute the bytes when writing on the wire as we are sure to release ByteBuf and don' want to compute twice.

Result:

More sane request computation