Update sse.rst
Doesn't it make sense first to show how to build a check and then explain how to set it? Also I see cancel check paragraph in the site, but I don't see it here, not sure if this is intended or I am missing something, also I must admit I am not quite familiar with how .rst work, so forgive me if I am talking nonsense and disregard. :)