Don't write absolute urls when not using HTTP/2
Motivation:

We're sending requests with absolute url when HTTP/2 is enabled but server only supports HTTP/1. This can cause server to break as such url is just a Netty hack.

Modification:

Only write absolute requests when actually using HTTP/2, not when HTTP/2 support is enabled.

Result:

Valid HTTP/1 requests when HTTP/2 is enabled