Have an inject built-in for incremental injection, close #3345
Motivation:
We would like to provide a builtin DSL element to create incremental workload and not be obliged to use map/flatMap to create them.

Modifications:
New DSL elements have been added for both open and closed workload to create incremental test.

Result:
It's now easier to create incremental tests with the DSL.