HTTP: resources computation doesn't account for data captured in main request with 3.0.0-RC1, close #3529
Motivation:
When session is modified in a request, the new session is not properly passed to the newResourceAggregatorForFetchedPage that build the resource request. Therefore, we don't have access to new session attributes in the resources block.

Modification:
The session that is passed to the newResourceAggregatorForFetchedPage is now the correct one.

Result:
We can now use session attributes computed in the request inside its resources block.