HTTP: resource fetching hangs when HTTP/2 is not enabled, close #3531
Motivation:
When HTTP/2 is not enabled, there are some part of the process where we inspect the prior knowledge map. This map exists only if HTTP/2 is enabled, and therefore, the code throws an exception in its absence. The exception is hidden, and make the resource fetching hangs infinitely.

Modification:
We no longer want to handle the absence of the map with an exception, but just by returning None meaning that we have no information about this remote in the prior knowledge map.

Result:
Run no longer hangs during resource fetching, and the isHttp2PriorKnowledge can be called even if HTTP/2 is not enabled.