JsonPath shouldn't escape `/`, close #3553
Motivation:

JSON spec requires / char to be escaped. In real world, no one does that as it breaks urls and this char doesn't conflict with anything else.

In 2.3, we were using Jackson for JSON escaping and it doesn't escape this char.

Modification:

Don't escape `/` in our own escaper introduced in Gatling 3.0.

Result:

Consistent JsonPath result with previous version.