Remove updates from polling, close #3564
Motivaion:

During some internals refactoring in Gatling 3, we've made it so that we usually no longer pile up session updates. The only exception is polling, where state is conciled when stopping.
Because of the refacting, we're capuring too much state. This increases the pace of running into OOME.

Modification:

As the piling up updates mechanism is very complicated and pooling only reconciles on stop, which is usually never done, let's just remove it. Polling can currently only be used for background requests with no sync with main flow.

Result:

No more mem leak with HTTP polling.