Feeder uniqueness should be based on identity, not equality, close #3585
Motivation:

We want to make sure we use the same FeederActor for the same feeder. Meaning that this uniqueness should be based on identity, not equality.
I suspect we once had standard classes, so equals wasn't overridden, but we at some point switched to case class and lost identity.

Modifications:

Use an IdentityHashMap to make sure we always use identity.

Result:

Feeder uniqueness proprly honored