Only perform retry when connection was pooled, close #3609
Motivation:

We perform retry on HTTP/1.1 when connection gets closed while request is flying.
We should only do so when the connection was obtained from the keep-alive pool because this retry should only be performed to deal with situation where we're reusing a connection and the remote server triggered keep-alive timeout on its end and client side hasn't be notified yet.

Modification:

* Store an attribute on the channel to flag it as pooled.
* only retry if this flag was set

Result:

Only retry on premature close on pooled connections