Handle SSE termination when channel was pooled, close #3638
Motivation:

We use `onTcpConnectSuccess` to keep track of the underlying channel so we can close it and notify the actor  that the stream was closed.
This doesn't work if the channel was obtained from the pool as `onTcpConnectSuccess` won't be called.

Modifications:

* Introduce `onWrite` to get notified of underlying channel.
* Refactor SseListener#close

Result:

SSE stream is properly closed when channel was obtained from the pool.