Only add Origin header when method isn't GET nor HEAD and Referer is present, close #3640
Motivation:

The way we forcefully generate Origin header is incorrect.

This header is used for CORS and forms.
Regarding the former, we expect user to provide it.
Regarding the latter, we can try to generate it from Referer.

Modification:

Only generate Origin when it's undefined and methods isn't GET nor HEAD and Referer is set (hence feature can be disabled along with autoReferer).

Result:

No more excessive Origin header generation.