Don't try to deduplicate resources, close #3786
Motivation:

We try to detect duplicate requests in resources, all the more as they can be aggregated from inferred and explicit resources.
Sadly, we only take url into account. This works fine for idempotent requests but not for mutating ones.

Modifications:

Let's drop this filter

Result:

Don't filter out requests that have the same url but are not identical, eg body.