Introduce "followedBy" for defining scenario sequential execution, close #3830
Motivation:

Lots of users try to have a way to execute scenarios sequentially.
The current workaround is to delay the injection profile with an empirical `nothingFor` step.

Modification:

Introduce `followedBy` to define children scenarios to be started once all users of parent scenario terminates.

Result:

It's now possible to define sequences of scenarios.