Introduce around assertion criterion, close #3845
Motivation:

It would be nice to have a convenience shortcup, based on `between` to define that actual value should be around a given target, plus or minus a given margin.

Modification:

Introduce `around`, that's just a shortcut that boils down to `between`.

Result:

Convenient way to define assertions with acceptance margin.