Set proxy-authorization header with Basic auth not set for clear HTTP, close #3877
Motivation:

Connection through an HTTP proxy for a clear HTTP request is handled directly, not with a HttpProxyHandler.
We don't set the proxy-authorization header manually.

Modification:

Add missing proxy-authorization request header when using an HTTP proxy and url is not secure.

Result:

proxy-authorization header properly set for clear HTTP requests