Introduce gatling-recorder config parameter 'recorder.http.useMethodAndUriAsPostfix'.
Motivation:
When using Gatling recorder, the produced report is not that easy to read for us, as the request names look quite alike having name request_$ID.

For us it is helpful, having also request method and URI in the request name; the produced reports are now quite useful to us.

Modifications:

Introduce gatling-recorder config parameter recorder.http.useMethodAndUriAsPostfix. When set to true, use the HTTP method and the URI as a postfix for http(s) requests name.

E.g. request name will be request_0:post_http://gatling.io/path1/file1 (instead of simply request_0). This improves legibility of recorded Gatling reports.

As this implementation sticked to approach of existing parameter recorder.http.useSimulationAsPrefix, it was quite a no-brainer.

Result:

Better request naming