Introduce ignoreDefaultHeaders, close  #3899
Motivation:

Users might want to define some dynamic headers on all requests but the first one of the scenario that would resolve it.

Modifications:

Introduce ignoreDefaultHeaders that make request only use the headers defined on the request and not the HttpProtocol.

Result:

Default headers can now be disabled on a specific request.