Account for HttpCheckScope when ordering HTTP checks, closes #3915
Motivation:

All checks are applied (best effort for saveAs) but only the first failure is reported.
This can give surprising result, depending on where the checks are defined, eg a request with a body check will report the body check failure even though status was 500.

Modification:

- Add priority attribute on HttpCheckScope
- Defined an ordering on HttpCheckScope based on the priority
- Sort checks based on scope priority, check origin and user defined order
- Add tests validating the desired behavior

Result:

Checks are sorted the following way:
  - Url > Status > Header > Chunks | Body > Time
  - Then request > protocol > default
  - Finally user defined order for request and protocol checks