Don't transform idempotent requests redirected with 301/302/303 into GET, close #3937
Motivation:

Browsers change POST (and similar mutating) requests into GET when redirecting on 301, 302 and 303.

We currently change method to GET as soon as the original method is not GET.

Modification:

Only switch to GET when original method is neither GET, nor HEAD, nor OPTIONS.

Result:

HEAD and OPTIONS requests are not transformed into GET after 301, 302 or 303 redirects.