Add new `noShard` option to not shard when running FrontLine clusters, close #3941
Motivation:

In Gatling 3.4.0, we're introducing andThen to have sequential scenarios.
There's a good chance some people will use it for single user initialisation, typically fetching an auth token.

We running with FrontLine, as we shard the load, some nodes wouldn't execute the scenario.
We need an option for not sharding, that would be a noop when not running with FrontLine.

Modifications:

Introduce `noShard` that disables sharding on injection profiles and scenario throttling.

Result:

Option to disable load sharding