Disable auto Referer when using shareConnections, close #3985
Motivation:

Sharing connections is for simulating server-to-server traffic where the client program uses a single connection pool.
In this case, we should enforce an extra `Referer` header that wouldn't be there in production (browsers only).

This is all the more an issue as server-to-server uses POST a lot, that triggers `Origin` header computation when there's a `Referer` one.

Modification:

Have `HttpProtocolBuilder#shareConnections` call `disableAutoReferer`.

Result:

No more extra `Referer` and `Origin` headers when simulating server-to-server traffic.