Recorded bodies paths are incorrect when no package is defined, close #4056
Motivation:

When package is empty, we generate an absolute path because we always concatenate with a "/" between the package and the classname.

This results in an absolute path that doesn't match the file's location in the classpath.

Modification:

Only add "/" separator when package is non empty.

Result:

Body files locations are properly generated when user omits the package.