Use LongRef for ThreadLocal
Motivation:

Using Long with ThreadLocal incurs boxing cost and
two visits to ThreadLocalMap for get and set.

Modifications:

Instead of using an immutable object and changing the reference in ThreadLocal,
an unchanging reference to a mutable LongRef is used.

Result:

In getAndSetLastUserId:
Even if userId is outside the range of LongCache (> 127), no new Long object is created.
Sinlge visit to ThreadLocalMap.