Clean up simulations and resources config options, close #4072
Motivation:

We have 2 options in gatling.conf named gatling.core.directory.simulations and gatling.core.directory.resources whose default values only make sense in the bundle:

for configuring the Recorder so that it knows where to generate files
for resolving resources files in Gatling => this is actually a workaround for a bug in gatling.sh|bat where user-files in put in the classpath instead of user-files/resources.

Modifications:

* move default values from gatling.conf to RecorderConfiguration
* fix gatling.sh|bat so that user-files/resources is in the classpath
* make custom location have higher precedence than classpath in Resource resolution to retain existing behavior

Result:

Consistent behavior across launchers: resources are in the classpath, with the possibility to override.