Introduce checkIf at request level that supports multiple checks, close #3806
Motivation:

`checkIf` currently lets the user configure a boolean condition to perform a single check.
For some use cases, it would be convenient to have a way to share the condition amongst multiple checks.

This cannot be achieved with current `checkIf` that has to return one single check to be passed along other checks to the `check` method. We don't want to lose this feature.

Modification:

We're introducing a new `checkIf` at the same level as `check` for HTTP and JMS protocols. This version supports multiple check parameters.
Beware of the syntax: you can only use curly braces when passing one single check. When passing multiple check, you have to use parens.

Result:

It's now possible to write

```
.checkIf(condition)(check1, check2)
```