RawFileBody and RawFileBodyPart with static paths shouldn't bypass gatling.core.rawFileBodiesInMemoryMaxSize, close #4134
Motivation:

We have distinct paths in RawFileBodies depending on the path being static or dynamic.
When the path is dynamic, we only cache the bytes when the payload is below a config limit.
When the path is static, we were bypassing this limit and aggressively loading the bytes, possibly resulting in OOME.

Modification:

Use the same caching strategy for both paths.

Result:

No more OOME when sending large files with static paths.