Fix OnConnectedChainEndAction, close #4167
Motivation:

OnConnectedChainEndAction is used to stash the action when reconnecting. In some cases, it's not properly executed.

Modification:

Fix wrong function where we were passing a `Session => Unit => Unit` to a `Session => Unit` so the actual work was never triggered when applying it.

Result:

Stashed action gets triggered when reconnecting.