Freeze Saxon version, see #4214
Since 11, Saxon has a hard dependency on xmlresolver. I couldn't find a way to bypass it, it the first thing happening in net.sf.saxon.Configuration.

Because of this move, it now pulls the following dependencies:

xmlresolver: 150259 B
xmlresolver/data classifier: 1027759 B
httpclient: 780321 B
httpcore: 328324 B
commons-logging: 61829 B
commons-codec: 353793 B
xml-apis: 109318 B
That's a grand total of 2811603 B, hence ~2.7MB of extra weight we don't want.
Moreover, we really don't want to ship and extra HTTP client than our internal one.

It try to report the issue to Saxonica but wasn't able to re-activate my account on their Developper forum.

I guess we're stuck with 10.6 and will have to live with it.