feat: Compile Java code with the --release flag when packaging for Gatling Enterprise Cloud
ref: MISC-379

Motivation:

Gatling Enterprise Cloud only supports Java 17. When running the bundle with a JDK 18+, we produce an invalid package.

Modification:

When running on Gatling Enterprise Cloud, detect the current Java version and compare to the max supported version from the server. If the version used is too high, use the javac --release flag to target the max supported version instead.