Fix Java feeder same instances detection, close #4240
Motivation:

We have a mechanism to ensure feeders used in multiple places work in a thread-safe fashion.
However, this mechanism is broken for Java because we detect the Scala wrapper instead.

Modification:

Use the underlying instance's identity hashcode.

Result:

Java feeder used in multiple places work in a tread-safe fashion.