feature: Introduce Gatling version check and deprecation warnings, close #4263
Motivation:

Some users tend to not upgrade their Gatling version, possibly because they're not aware of the latest release and the new features and bug fixes.
Let's advertise the new releases in the console and add a deprecation warning in the HTML reports when the version is more than 1 year old.

Result:

Gatling now invites users to upgrade.