Don't display a stracktrace when there's no simulation, close #4272
Motivation:

We display a long stacktrace when the bundle doesn't contain any Simulation.
We should just display the error message.

Modification:

Catch Exception based on message. This is very sad, but we have no choice as the actual error is the forked JVM exiting, so the Exception is an ExecuteException and we can't use our own Exception type.

Result:

No more ugly stacktrace in this case