Enforce bytecode version in Package command, close #4278
Motivation:

Java 18 users (or users not aware this is a STS version and not LTS) might want to generate the package and upload it manually. If so, we must enforce Java 17.

Bytecode version is only limited on the CLI direct upload. It should be enforced on the package generation too.

Modification:

Try to fetch max supported version from the API. If it's not available, fall back to Java 17.

Result:

Java 18 users can generate Gatling Enterprise Cloud compatible packages.