Fix Response Time Ranges labels in console, close #4302
Motivation:

Gatling 3.8 changed one of the labels with a `<br>` tag so it would be displayed on 2 lines.
This label is currently shared with the console so this HTML tag appears here too.

Modification:

Have 2 different labels: one for the console and one for the HTML reports.

Result:

No more `<br>` tag in the console.