adds Nullable annotation to Session methods' return type
Motivation:

Kotlin has nullable types, but it needs annotation from Java code to know if something is nullable.

Modifications:

Return types in the Session API are annotated @Nullable.
Comments in the Kotlin sample use the types as they are called in Kotlin.

Result:

Kotlin compiler now knows the nullability of values.