Encode url even when it's a static url without queryParams, close #4366
Motivation:

First, it's not a regression introduced in 3.9.0. gatling has been behaving this way for ages, expecting that when the user passes a simple static url, he makes the effort of passing a valid url and not one that's not properly encoded.

Even if I think this expectation makes sense (it took many years for this ticket to get raised), we can perform encoding in this case without performance penalty. So let's do it.

Modification:

Call `UriEncoder` also on the "static url + signle baseUrl + no queryParams" shortcut path