Make multiple feed generate either j.u.List or Scala Seq instead of Array, close #4423
Motivation:

The current behavior of generating arrays is not consistent with the rest of Gatling where the Java side deals with java.util.List while the Scala one deals with Seq
Arrays are painful to deal with at runtime with reflection
Let's change the behavior.

This would fix scenarios where foreach crashes at runtime with a CCE as it expects a Seq.