fix: Ws and SSE check timeout cancellation, close #4434
Motivation:

We have a bug where we always cancel the current timeout as soon as we receive a matching message. But the sequence could have more than one check.

Motivation:

Only cancel the timeout when the check fail or the sequence is now complete.

Result:

Timeout works when using check sequences with more than one check.