chore: freeze activemq major version because of javax to jakarta namespace change
Motivation:

Oracle has decided to donate JEE to the community but required that it moves away from the `javax` namespace. As a result, new versions use the `jakarta` namespace.

ActiveMQ 6 has done the move so it's not compatible with our existing JMS support.

See https://github.com/gatling/gatling/issues/4443