Doc: Update Frontline Configuration and Upgrade Cassandra to 4.1 for Consistency Across Installations (#850)
Modifies the Frontline configuration to align with the contact points reference used in Cassandra 4.1's Datastax Java driver. This update is necessary because Cassandra 4.1 is referenced in the sample YAML resource.

Upgrade Cassandra Docker Installation to Version 4.1

Upgrades Cassandra from version 4.0 to 4.1 in the Docker installation process for consistency with the Kubernetes installation, which uses Cassandra 4.1.