feat: Introduce support for PROXY protocol, close #4518
Motivation:

Some applications might require a fake PROXY protocol header to perform properly based on the information in there, in particular the source address and port.

https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt

Modifications:

Introduce `proxyProtocolSourceIpV4Address` and `proxyProtocolSourceIpV6Address` on HttpProtocol.

When one of these functions are defined AND a proxy is defined AND the the request host isn't in the "no-proxy-for" scope, Gatling will generate fake PROXY protocol headers based on the fake address returned by these functions.

Note that Gatling will resolve the fonction for every request execution so if a virtual user must keep the same address for his whole lifetime, the value returned by the function must always be the same, eg pull from a feeder.
Note: