fix: truncate error messages so they don't overflow the buffer, close #4660
Motivation:

Error messages that are longer than 8kB cause a buffer overflow in BufferedChannelFileWriter.

Anyway, it doesn't make sense to use error messages that are so long: it's most likely a user error and we can't display them in the reports or the console.

Modification:

Truncate error messages to 200 characters.